changed helix-query and moved some js to utils
